requirements.txt
hist
mplhep
matplotlib
numpy
scipy
uproot
scikit-learn
boost-histogram
awkward1
tqdm
